Removed CORSFilter, refactored config, controller, repository, service
