+++ Emp-Management-system/src/main/java/com/ems/config/JwtOncePerRequestAuthFilter.java
@@ -1,0 +1,102 @@
+package com.ems.config;
+
+import java.io.IOException;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+import io.jsonwebtoken.ExpiredJwtException;
+import io.jsonwebtoken.SignatureException;
+import jakarta.annotation.Resource;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
+/**
+ * <li> A Security Filter that <b>Authenticates</b> each and every Request using the <b>JWT</b> token from Request Authorization Header. <br></li>
+ * 
+ * <li> After authentication, a Authentication Object is generated and added to Security Context.</li>
+ * 
+ * <li> <i>HEADER_STRING</i> and <i>TOKEN_PREFIX</i> are configured in application.properties</li>
+ *
+ */
+public class JwtOncePerRequestAuthFilter extends OncePerRequestFilter {
+
+	@Value("${jwt.header.string}")
+	public String HEADER_STRING;
+
+	@Value("${jwt.token.prefix}")
+	public String TOKEN_PREFIX;
+
+	@Resource(name = "userService")
+	private UserDetailsService userDetailsService;
+
+	@Autowired
+	private TokenUtil jwtTokenUtil;
+
+	/**
+	 * <b>Authenticate</b> the user with JWT token and the Stored UserDetails
+	 */
+	@Override
+	protected void doFilterInternal(HttpServletRequest req, HttpServletResponse res, FilterChain chain)
+			throws IOException, ServletException {
+
+		String header = req.getHeader(HEADER_STRING);
+		
+		if (header == null) {
+			logger.warn("Token Header is null");
+		}
+
+		String username = null;
+		String authToken = null;
+
+		// Extract the username from the JWT token
+		if (header != null && header.startsWith(TOKEN_PREFIX)) {
+
+			authToken = header.replace(TOKEN_PREFIX, "");
+
+			try {
+				username = jwtTokenUtil.getUsernameFromToken(authToken);
+			} catch (IllegalArgumentException e) {
+				logger.error("Error occurred while retrieving Username from Token", e);
+			} catch (ExpiredJwtException e) {
+				logger.warn("The token has expired", e);
+			} catch (SignatureException e) {
+				logger.error("Authentication Failed. Invalid username or password.");
+			}
+
+		} else {
+			
+			logger.warn("Bearer string not found, ignoring the header");
+		}
+
+		// Authenticate user with username from JWT and the UserDetails with that username.
+		if (username != null && SecurityContextHolder.getContext().getAuthentication() == null) {
+			
+			UserDetails userDetails = userDetailsService.loadUserByUsername(username);
+			
+			logger.info("Username from UserDetails: " + userDetails.getUsername());
+			
+			if (jwtTokenUtil.validateToken(authToken, userDetails)) {
+				
+				UsernamePasswordAuthenticationToken authentication = jwtTokenUtil.getAuthenticationToken(authToken,
+						SecurityContextHolder.getContext().getAuthentication(), userDetails);
+				
+				authentication.setDetails(new WebAuthenticationDetailsSource().buildDetails(req));
+								
+				SecurityContextHolder.getContext().setAuthentication(authentication);
+				
+				logger.info("User authenticated: " + username + ", setting security context");
+			}
+		}
+
+		chain.doFilter(req, res);
+	}
+}
