+++ Emp-Management-system/src/main/java/com/ems/config/ModelMapperConfiguration.java
@@ -4,8 +4,13 @@ import org.modelmapper.ModelMapper;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
+/**
+ * A Configuration class to provide <b>ModelMapper</b> Instance.
+ * 
+ * The instance will be used to map objects between Entity and DTO.
+ */
 @Configuration
-public class AppConfig {
+public class ModelMapperConfiguration {
 
     @Bean
     public ModelMapper modelMapper() {
