+++ Emp-Management-system/src/main/java/com/ems/config/PasswordEncoderConfiguration.java
@@ -4,11 +4,15 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
+/**
+ * Provides BCryptPasswordEncoder for hashing passwords 
+ */
 @Configuration
-public class PasswordEncoder {
+public class PasswordEncoderConfiguration {
 
     @Bean
     public BCryptPasswordEncoder encoder(){
         return new BCryptPasswordEncoder();
     }
+    
 }
