--- Emp-Management-system/src/main/java/com/ems/config/UnauthorizedEntryPoint.java
@@ -1,27 +1,0 @@
-package com.ems.config;
-
-import java.io.IOException;
-import java.io.Serializable;
-
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-
-@Component
-public class UnauthorizedEntryPoint implements AuthenticationEntryPoint, Serializable {
-
-    /**
-	 * 
-	 */
-	private static final long serialVersionUID = 1L;
-
-	@Override
-    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException {
-        response.sendError(HttpServletResponse.SC_UNAUTHORIZED, "Unauthenticated");
-    }
-
-}
