+++ Emp-Management-system/src/main/java/com/ems/model/AuthToken.java
@@ -1,46 +1,22 @@
 package com.ems.model;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 /**
  * Represents an authentication token.
  */
 @Data
+@AllArgsConstructor
+@NoArgsConstructor
 public class AuthToken {
+	
     private String token;
-
     private long id;
-    /**
-     * Constructs a new AuthToken object.
-     */
-    public AuthToken() {
-    }
-
-    /**
-     * Constructs a new AuthToken object with the specified token.
-     * 
-     * @param token the authentication token
-     */
+    
     public AuthToken(String token) {
-        this.token = token;
+    	this.token = token;
     }
 
-    /**
-     * Returns the authentication token.
-     * 
-     * @return the authentication token
-     */
-    public String getToken() {
-        return token;
-    }
-
-    /**
-     * Sets the authentication token.
-     * 
-     * @param token the authentication token to be set
-     */
-    public void setToken(String token,long id) {
-        this.token = token;
-        this.id=id;
-    }
 }
