--- Emp-Management-system/src/main/java/com/ems/model/File.java
@@ -1,50 +1,0 @@
-package com.ems.model;
-
-
-import com.fasterxml.jackson.annotation.JsonBackReference;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "files")
-@Builder
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class File {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String name;
-    private String fileUrl;
-    
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    @JsonBackReference  
-    private User user;
-
-    
-    public File(User user, String name, String fileUrl) {
-        this.user = user;
-        this.name = name;
-        this.fileUrl = fileUrl;
-        
-    }
-
-    @Override
-    public String toString() {
-        return "UserFile [documentId=" + id + ", fileName=" + name + ", fileUrl=" + fileUrl + "]";
-    }
-}
