+++ Emp-Management-system/src/main/java/com/ems/model/LoginUser.java
@@ -8,8 +8,8 @@ import lombok.Setter;
 @Getter
 @Setter
 public class LoginUser {
+	
     private String email;
     private String password;
-
     
 }
