+++ Emp-Management-system/src/main/java/com/ems/model/dto/UserFileDTO.java
@@ -1,10 +1,12 @@
-package com.ems.model;
+package com.ems.model.dto;
 
 import lombok.Data;
 
 @Data
 public class UserFileDTO {
+	
 	private int documentId;
     private String fileName;
     private String fileUrl;
+    
 }
