+++ Emp-Management-system/src/main/java/com/ems/repository/AccessHistoryDao.java
@@ -2,7 +2,7 @@ package com.ems.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.ems.model.AccessHistory;
+import com.ems.model.Entity.AccessHistory;
 
 public interface AccessHistoryDao extends JpaRepository<AccessHistory, Long>{
