+++ Emp-Management-system/src/main/java/com/ems/repository/FileDao.java
@@ -2,7 +2,8 @@ package com.ems.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.ems.model.File;
+import com.ems.model.Entity.File;
 
 public interface FileDao extends JpaRepository<File, Long> {
+	
 }
