+++ Emp-Management-system/src/main/java/com/ems/repository/MessageDao.java
@@ -2,8 +2,7 @@ package com.ems.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.ems.model.Message;
-
+import com.ems.model.Entity.Message;
 
 public interface MessageDao extends JpaRepository<Message, Long> {
