+++ Emp-Management-system/src/main/java/com/ems/repository/RoleDao.java
@@ -1,17 +1,17 @@
 package com.ems.repository;
 
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import com.ems.model.Role;
+import com.ems.model.Entity.Role;
 
 import jakarta.transaction.Transactional;
 
 @Repository
 public interface RoleDao extends JpaRepository<Role, Long> {
+	
     Role findRoleByName(String name);
     
     
@@ -19,6 +19,8 @@ public interface RoleDao extends JpaRepository<Role, Long> {
     @Transactional
     @Query(value = "INSERT INTO user_roles (user_id, role_id) " +
                    "SELECT :userId, 1 WHERE NOT EXISTS " +
-                   "(SELECT 1 FROM user_roles WHERE role_id = 1 AND user_id = :userId)", nativeQuery = true)
-    void makeUserAdmin(Long userId);
+                   "(SELECT 1 FROM user_roles WHERE role_id = 1 AND user_id = :userId)", 
+           nativeQuery = true)
+	void makeUserAdmin(Long userId);
+    
 }
