+++ Emp-Management-system/src/main/java/com/ems/repository/UserDao.java
@@ -8,21 +8,19 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import com.ems.model.User;
-import com.ems.model.UserStatus;
+import com.ems.model.Entity.User;
+import com.ems.util.UserStatus;
 
 @Repository
 public interface UserDao extends JpaRepository<User, Long> {
-	//Optional<User> findByEmail(String email);
 
-	 @Query("SELECT u FROM User u JOIN u.roles r WHERE r.name = 'ADMIN'")
-	    List<User> findAllAdmins();
-	 
-	 
-	 @Query("SELECT u FROM User u WHERE u.status = :status")
-		List<User> findByStatus(@Param("status") UserStatus status);
-	 
-	 Optional<User> findByEmail(String email);
-	 
-	  int countByStatus(UserStatus status);
+	@Query("SELECT u FROM User u JOIN u.roles r WHERE r.name = 'ADMIN'")
+	List<User> findAllAdmins();
+
+	@Query("SELECT u FROM User u WHERE u.status = :status")
+	List<User> findByStatus(@Param("status") UserStatus status);
+
+	Optional<User> findByEmail(String email);
+
+	int countByStatus(UserStatus status);
 }
