+++ Emp-Management-system/src/main/java/com/ems/service/FileService.java
@@ -2,7 +2,8 @@ package com.ems.service;
 
 import org.springframework.web.multipart.MultipartFile;
 
-public interface IFileService {
+public interface FileService {
 
 	String uploadFile(MultipartFile file, Long userId);
+	
 }
