+++ Emp-Management-system/src/main/java/com/ems/service/RoleService.java
@@ -1,11 +1,9 @@
 package com.ems.service;
 
-import com.ems.model.Role;
+import com.ems.model.Entity.Role;
 
-//Importing the Role model
-
-//Declaring the RoleService interface
 public interface RoleService {
-	// Method to find a Role by its name
+	
 	Role findByName(String name);
+	
 }
