+++ Emp-Management-system/src/main/java/com/ems/service/UserService.java
@@ -4,43 +4,28 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.web.multipart.MultipartFile;
 
-import com.ems.model.User;
-import com.ems.model.UserDto;
-
-
+import com.ems.model.Entity.User;
+import com.ems.model.dto.UserDto;
 
 public interface UserService {
 
-	// Saves a user
 	User save(UserDto user);
-
-	// Retrieves all users
-	List<User> findAll();
-
-	// Retrieves a user by username
-//    User findOne(String email);
 
 	User createUser(UserDto user);
 
-	//String sendHtmlEmail(String toEmail, String subject, String body);
+	User updateUser(Long userId, UserDto updatedUserDto);
 
-	public UserDetails loadUserByUsername(String email);
-	
 	List<User> findAllAdmins();
-	
-	
-	//public List<User> uploadAndCreateUsers(MultipartFile file);
-	
+
+	List<User> findAll();
+
+	UserDetails loadUserByUsername(String email);
+
 	Optional<User> findByEmail(String email);
 
-	
-	public User updateUser(Long userId, UserDto updatedUserDto);
-
 	Optional<User> findByid(Long id);
-	
+
 	public void makeUserAdmin(Long userId);
-	
-	
+
 }
