+++ Emp-Management-system/src/main/java/com/ems/service/serviceImpl/AccessHistoryServiceImpl.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.ems.model.AccessHistory;
+import com.ems.model.Entity.AccessHistory;
 import com.ems.repository.AccessHistoryDao;
