+++ Emp-Management-system/src/main/java/com/ems/service/serviceImpl/FileServiceImpl.java
@@ -18,10 +18,10 @@ import com.amazonaws.regions.Regions;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import com.amazonaws.services.s3.model.PutObjectRequest;
-import com.ems.service.IFileService;
+import com.ems.service.FileService;
 
 @Service
-public class FileServiceImpl implements IFileService {
+public class FileServiceImpl implements FileService {
 
     private static final Logger logger = LoggerFactory.getLogger(FileServiceImpl.class);
