+++ Emp-Management-system/src/main/java/com/ems/service/serviceImpl/RoleServiceImpl.java
@@ -4,7 +4,7 @@ package com.ems.service.serviceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.ems.model.Role;
+import com.ems.model.Entity.Role;
 import com.ems.repository.RoleDao;
 import com.ems.service.RoleService;
