+++ Emp-Management-system/src/main/java/com/ems/util/Role.java
@@ -1,0 +1,11 @@
+package com.ems.util;
+
+public enum Role {
+	ADMIN,
+	USER;
+	
+	@Override
+	public String toString() {
+		return name();
+	}
+}
