+++ Emp-Management-system/src/main/resources/application.properties
@@ -1,16 +1,20 @@
 spring.application.name=Emp-Management-system
 
 
+# JWT Authentication Configuration ------------------------------------------------
 
 jwt.token.validity=50000
 jwt.signing.key=YourSignInKey
+	
 jwt.authorities.key=roles
 jwt.token.prefix=Bearer
 jwt.header.string=Authorization
 
+# Database Configuration ----------------------------------------------------------
+
 spring.datasource.url=jdbc:mysql://localhost:3306/ems?createDatabaseIfNotExist=true
 spring.datasource.username=root
-spring.datasource.password=root@123
+spring.datasource.password=123
 
 spring.jpa.show-sql=true
 spring.jpa.hibernate.ddl-auto=update
@@ -18,11 +22,13 @@ spring.user.datasource.driver-class-name=com.mysql.jdbc.Driver
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
 server.port=8999
 
+# Logging --------------------------------------------------------------------------
+
 #logging.level.root=DEBUG
+#logging.level.org.springframework.security=TRACE
+#logging.level.root=ERROR
 
-#TO insert Role
-
-#Aws credientials
+# AWS Credentials ------------------------------------------------------------------
 
 aws.accessKeyId=AKIATCKATLCFZ3WR5RPT
 aws.secretAccessKey=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
@@ -30,13 +36,12 @@ aws.s3.bucketName=ems-bucket-13-02
 aws.s3.region=ap-south-1
 
 
-#sending mail
+# Mail Configuration ----------------------------------------------------------------
+
 spring.mail.host=smtp.gmail.com
 spring.mail.port=587
 spring.mail.username=abhishekbhosale676@gmail.com
 spring.mail.password=byxr ohjy mlrn oimn
-
-
 
 spring.servlet.multipart.max-file-size=20MB
 spring.servlet.multipart.max-request-size=20MB
